+++ scripts/triggers-end-to-end-tests/tests.ts
@@ -22,7 +22,7 @@ const ADMIN_CREDENTIAL = {
  */
 const TEST_SETUP_TIMEOUT = 80000;
 const EMULATORS_WRITE_DELAY_MS = 5000;
-const EMULATORS_SHUTDOWN_DELAY_MS = 5000;
+const EMULATORS_SHUTDOWN_DELAY_MS = 7000;
 const EMULATOR_TEST_TIMEOUT = EMULATORS_WRITE_DELAY_MS * 2;
 
 /*
@@ -428,5 +428,71 @@ describe("function triggers", () => {
     this.timeout(TEST_SETUP_TIMEOUT);
     const v2response = await test.invokeHttpFunction("onreqv2timeout");
     expect(v2response.status).to.equal(500);
+  });
+
+  describe("disable/enableBackgroundTriggers", () => {
+    before(() => {
+      test.resetCounts();
+    });
+
+    it("should disable background triggers", async function (this) {
+      this.timeout(TEST_SETUP_TIMEOUT);
+
+      const response = await test.disableBackgroundTriggers();
+      expect(response.status).to.equal(200);
+
+      await new Promise((resolve) => setTimeout(resolve, EMULATORS_WRITE_DELAY_MS));
+
+      await Promise.all([
+        // TODO(danielylee): Trying to respond to all triggers at once often results in Functions
+        // Emulator hanging indefinitely. Only triggering 1 trigger for now. Re-enable other triggers
+        // once the root cause is identified.
+        // test.writeToRtdb(),
+        // test.writeToFirestore(),
+        // test.writeToPubsub(),
+        // test.writeToDefaultStorage(),
+        test.writeToAuth(),
+      ]);
+
+      await new Promise((resolve) => setTimeout(resolve, EMULATORS_WRITE_DELAY_MS * 2));
+
+      // expect(test.rtdbTriggerCount).to.equal(0);
+      // expect(test.rtdbV2TriggerCount).to.eq(0);
+      // expect(test.firestoreTriggerCount).to.equal(0);
+      // expect(test.pubsubTriggerCount).to.equal(0);
+      // expect(test.pubsubV2TriggerCount).to.equal(0);
+      expect(test.authTriggerCount).to.equal(0);
+    });
+
+    it("should re-enable background triggers", async function (this) {
+      this.timeout(TEST_SETUP_TIMEOUT);
+
+      const response = await test.enableBackgroundTriggers();
+      expect(response.status).to.equal(200);
+
+      await new Promise((resolve) => setTimeout(resolve, EMULATORS_WRITE_DELAY_MS));
+
+      await Promise.all([
+        // TODO(danielylee): Trying to respond to all triggers at once often results in Functions
+        // Emulator hanging indefinitely. Only triggering 1 trigger for now. Re-enable other triggers
+        // once the root cause is identified.
+        // test.writeToRtdb(),
+        // test.writeToFirestore(),
+        // test.writeToPubsub(),
+        // test.writeToDefaultStorage(),
+        test.writeToAuth(),
+      ]);
+
+      await new Promise((resolve) => setTimeout(resolve, EMULATORS_WRITE_DELAY_MS * 3));
+      // TODO(danielylee): Trying to respond to all triggers at once often results in Functions
+      // Emulator hanging indefinitely. Only triggering 1 trigger for now. Re-enable other triggers
+      // once the root cause is identified.
+      // expect(test.rtdbTriggerCount).to.equal(1);
+      // expect(test.rtdbV2TriggerCount).to.eq(1);
+      // expect(test.firestoreTriggerCount).to.equal(1);
+      // expect(test.pubsubTriggerCount).to.equal(1);
+      // expect(test.pubsubV2TriggerCount).to.equal(1);
+      expect(test.authTriggerCount).to.equal(1);
+    });
   });
 });