+++ src/frameworks/nuxt/index.ts
@@ -3,10 +3,13 @@ import { readFile } from "fs/promises";
 import { join } from "path";
 import { gte } from "semver";
 import { findDependency, FrameworkType, relativeRequire, SupportLevel } from "..";
+import { warnIfCustomBuildScript } from "../utils";
 
 export const name = "Nuxt";
 export const support = SupportLevel.Experimental;
 export const type = FrameworkType.Toolchain;
+
+const DEFAULT_BUILD_SCRIPT = ["nuxt build"];
 
 export async function discover(dir: string) {
   if (!(await pathExists(join(dir, "package.json")))) return;
@@ -27,6 +30,9 @@ export async function discover(dir: string) {
 export async function build(root: string) {
   const { buildNuxt } = await relativeRequire(root, "@nuxt/kit");
   const nuxtApp = await getNuxt3App(root);
+
+  await warnIfCustomBuildScript(root, name, DEFAULT_BUILD_SCRIPT);
+
   await buildNuxt(nuxtApp);
   return { wantsBackend: true };
 }