+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-- Add support for object list using certain Admin SDKs (#5208)
-- Add support for deploying Next.js applications that utilize middleware
-- Add basic support for deploying Nuxt 2+3 applications (#5321)
+- Add sharp NPM module to Cloud Functions when using Next.js Image Optimization (#5238)
+- Adds user-defined env vars into the functions emulator (#5330).
+- Support Next.js Middleware (#5320)
+- Log the reason for a Cloud Function if needed in Next.js (#5320)